Model review

+++ analysis/R/functions.R
@@ -10,7 +10,7 @@ setup_schema <- schema(
   `resources-dist` = utf8(), `proposal-sigma` = float64(), 
   `n-teams` = int64(), `third-party-funding-ratio` = int64(), 
   `utility-change` = float64(), b_utility = int64(), network = utf8(), 
-  `funded-share` = int64(),`data-sharing?` = bool(), 
+  `funded-share` = float64(),`data-sharing?` = bool(), 
   `max-initial-utility` = int64(), `[step]` = int64(), 
   `gini [resources] of teams` = float64(), `gini [total-funding] of teams` = float64(), 
   `mean [effort] of teams` = float64(), `%-sharing` = int64(),
